fixed deploy and added real discount in sidebar
