+++ app/components/order-sidebar/order-product-card/order-product-card.tsx
@@ -1,3 +1,4 @@
+'use client'
 import Image from "next/image";
 import { useState } from "react";
